:sparkles: Add wallet support to web3 object
